Fix emoji handling
Emoji and some other characters can be a single unicode character with a
value that cannot be encoded on 16 bits. In those cases, JavaScript
consider them as 2 UTF-16 code units: it is called a surrogate pair.
Prosemirror is a JavaScript library, and the go port has to count the
from/to for the steps in the same way to be compatible. It means that we
have to transform the UTF-8 strings in slices of UTF-16 code units in a
few places of the code, which is not too difficult as Go stdlib helps
us a lot.